i2c: bma250: gyro wake: fixing a condition when gyro is left enabled
Sometimes gyro off from userspace is happening BEFORE actual mipi
panel power off. But while screen is On, gyro wake is skipping the
disabling of the sensor. Now we will store the time of the last
such skipping of gyro disable try, and now will check if the
panel power off is very close to it in time, and switch the sensor
off if it is close enough. This prevents unneeded on time and
wakelocks.